Implement RedisCluster and CloudMemorystoreInstance controllers
I've experimented with a new pattern for the CloudMemorystore controller. My
goal was to achieve the testability of the existing blessed pattern (i.e.
replacable connect, create, sync, and delete methods), but with a stronger
decoupling of the components we'd like to test separately.

In my design Reconciler takes a connecter interface that knows how to return a
createsyncdeleter, i.e. an interface that can create, sync, and delete instances
with some external store. We instantiate a new createsyncdeleter on each
reconcile loop, allowing us to potentially use different credentials each run.

I also removed use of the Kubernetes clientset, which was used a few places in
other controllers despite (as best I can tell) being functionally equivalent to
the controller runtime client.Client.

Signed-off-by: Nic Cope <negz@rk0n.org>