Refactor ReferenceTo to take an explicit Group, Version, and Kind.
We don't seem to be able to rely on Go objects retaining type metadata
when deserialised. I've read (but not confirmed) that the decoding process
strips out TypeMeta, and observed TypeMeta disappearing during an Update()
call.

This API mirrors that of metav1.NewControllerRef. The caller can still attempt
to use TypeMeta if it is known to exist, i.e.:

```go
ref := meta.ReferenceTo(o, o.GetObjectKind().GroupVersionKind())
```

Signed-off-by: Nic Cope <negz@rk0n.org>